+++ Minisuper.API/MinisuperZeus.BW/Interfaces/BW/IGestionarListaDeDeseosBW.cs
@@ -16,7 +16,7 @@ namespace MinisuperZeus.BW.Interfaces.BW
 
         public Task<bool> EliminarDeLaLista(int idProducto);
 
-        public Task<decimal> TotalAPagarConIVA(decimal impuesto);
+        public Task<float> TotalAPagarConIVA(float impuesto);
 
         public Task<bool> AgregarCantidadProducto(int idProducto, int cantidad);
